Update test task to sync it with the second stage
