+++ src/index.js
@@ -1,3 +1,12 @@
-// Please run your solution from this file
+import {prepareData, renderData} from './solution'
 
-console.log("Hello from %csrc/index.js", "font-weight:bold");
+const filterParams = {
+  year: 2018,
+  customerName: 'NASA',
+}
+
+// example of run, you could leave it or modify however you want
+fetch('https://api.spacexdata.com/v3/launches/past')
+  .then(response => response.json())
+  .then(prepareData(filterParams))
+  .then(renderData)
