+++ src/solution.js
@@ -1,8 +1,4 @@
 // Please implement your solution in this file
 
-const noop = () => {};
-
-module.exports = {
-  prepareData: noop,
-  renderData: noop
-};
+export const prepareData = () => {}
+export const renderData = () => {}
